ComBat v3.34 batch correction Older version of ComBat (sva version 3.34.0). This is for compatibility with the replication of SCoPE2. ComBat allows users to adjust for batch effects in datasets where the batch covariate is known, using methodology described in Johnson et al. 2007. It uses either parametric or non-parametric empirical Bayes frameworks for adjusting data for batch effects. Users are returned an expression matrix that has been corrected for batch effects. The input data are assumed to be cleaned and normalized before batch effect removal.

ComBat v3.34 batch correction

Older version of ComBat (sva version 3.34.0). This is for compatibility with the replication of SCoPE2.

ComBat allows users to adjust for batch effects in datasets where the batch covariate is known, using methodology described in Johnson et al. 2007. It uses either parametric or non-parametric empirical Bayes frameworks for adjusting data for batch effects. Users are returned an expression matrix that has been corrected for batch effects. The input data are assumed to be cleaned and normalized before batch effect removal.

ComBatv3.34(
  dat,
  batch,
  mod = NULL,
  par.prior = TRUE,
  prior.plots = FALSE,
  mean.only = FALSE,
  ref.batch = NULL,
  BPPARAM = bpparam("SerialParam")
)

Arguments

dat
Genomic measure matrix (dimensions probe x sample) - for example, expression matrix

batch
Batch covariate (only one batch allowed)

mod
Model matrix for outcome of interest and other covariates besides batch

par.prior
(Optional) TRUE indicates parametric adjustments will be used, FALSE indicates non-parametric adjustments will be used

prior.plots
(Optional) TRUE give prior plots with black as a kernel estimate of the empirical batch effect density and red as the parametric

mean.only
(Optional) FALSE If TRUE ComBat only corrects the mean of the batch effect (no scale adjustment)

ref.batch
(Optional) NULL If given, will use the selected batch as a reference for batch adjustment.

BPPARAM
(Optional) BiocParallelParam for parallel operation

Value

data A probe x sample genomic measure matrix, adjusted for batch effects.

Examples

library(bladderbatch)

#> Error in library(bladderbatch): there is no package called ‘bladderbatch’

data(bladderdata)

#> Warning: data set ‘bladderdata’ not found

dat <- bladderEset[1:50,]

#> Error in eval(expr, envir, enclos): object 'bladderEset' not found

pheno = pData(dat)

#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'pData': object 'dat' not found

edata = exprs(dat)

#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'exprs': object 'dat' not found

batch = pheno$batch

#> Error in eval(expr, envir, enclos): object 'pheno' not found

mod = model.matrix(~as.factor(cancer), data=pheno)

#> Error in terms.formula(object, data = data): object 'pheno' not found

# parametric adjustment
combat_edata1 = ComBat(dat=edata, batch=batch, mod=NULL, par.prior=TRUE, prior.plots=FALSE)

#> Error in ComBat(dat = edata, batch = batch, mod = NULL, par.prior = TRUE,     prior.plots = FALSE): could not find function "ComBat"

# non-parametric adjustment, mean-only version
combat_edata2 = ComBat(dat=edata, batch=batch, mod=NULL, par.prior=FALSE, mean.only=TRUE)

#> Error in ComBat(dat = edata, batch = batch, mod = NULL, par.prior = FALSE,     mean.only = TRUE): could not find function "ComBat"

# reference-batch version, with covariates
combat_edata3 = ComBat(dat=edata, batch=batch, mod=mod, par.prior=TRUE, ref.batch=3)

#> Error in ComBat(dat = edata, batch = batch, mod = mod, par.prior = TRUE,     ref.batch = 3): could not find function "ComBat"